# static/swagger.yaml
openapi: 3.0.0
info:
  title: Booking API
  description: API to manage bookings in a barber shop
  version: 1.0.0

servers:
  - url: http://localhost:5000

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

security:
  - bearerAuth: []

paths:
  /auth/register:
    post:
      summary: Register a new user
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - password
                - birthDate
              properties:
                email:
                  type: string
                  example: john@email.com
                password:
                  type: string
                  example: admin123
      responses:
        '201':
          description: User registered successfully
        '400':
          description: User already exists

  /auth/login:
    post:
      summary: Login user and get JWT token
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  example: john@email.com
                password:
                  type: string
                  example: admin123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                        example: d6957f05-117f-4d4e-bb22-a98512898ef1
                      name:
                        type: string
                        example: John Doe
                      email:
                        type: string
                        example: johndoe@emial.com
                      birthDate:
                        type: string
                        example: Tue, 10 Mar 2025 00:00:00 GMT
                      phone:
                        type: string
                        example: +55 79 98877-6655
                      createdAt:
                        type: string
                        example: Wed, 21 May 2025 00:00:00 GMT
                      updatedAt:
                        type: string
                        example: Fri, 23 May 2025 00:00:00 GMT
                      loyaltyPackage:
                        type: string
                        example: monthly
                      avaliableServicesNumber:
                        type: integer
                        example: 0
                      flStatus:
                        type: string
                        example: A
        '401':
          description: Invalid credentials

  /auth/validate-token:
    get:
      summary: Validates the JWT token
      tags:
        - Authentication
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: Valid token
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Valid token
                  user:
                    type: object
                    properties:
                      email:
                        type: string
                        example: johndoe@email.com
                      user_id:
                        type: string
                        example: d6957f05-117f-4d4e-bb22-a98512898ef1
        '401':
          description: Invalid, expired or not provided token

  /auth/request-password-reset:
    post:
      summary: Request password reset
      tags:
        - Authentication
      description: Sends an email to the user with a password reset link or token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
      responses:
        '200':
          description: Password reset email sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Password reset email sent successfully
        '404':
          description: User not found
        '500':
          description: Internal server error

    /auth/reset-password/{token}:
      post:
        summary: Reset password using token
        tags:
          - Auth
        description: Reset the user password using the provided token
        parameters:
          - name: token
            in: path
            required: true
            description: Password reset token sent via email
            schema:
              type: string
        requestBody:
          required: true
          content:
            application/json:
              schema:
                type: object
                properties:
                  newPassword:
                    type: string
                    example: "NewStrongPassword123"
        responses:
          '200':
            description: Password reset successfully
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    message:
                      type: string
                      example: Password reset successfully
          '400':
            description: Invalid or expired token
          '500':
            description: Internal server error

  /auth/reset-password/{token}:
    post:
      summary: Reset user password
      tags:
        - Authentication
      parameters:
        - in: path
          name: token
          required: true
          schema:
            type: string
            example: "d6957f05-117f-4d4e-bb22-a98512898ef1"
          description: Token for password reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                newPassword:
                  type: string
                  example: "NewPassword123!"
              required:
                - newPassword
                - confirmPassword
      responses:
        '200':
          description: Password reset successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Password reset successfully"
        '404':
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User not found"
        '500':
          description: Internal server error

  /booking/insert/{id_user}:
    post:
      summary: Create a new booking
      tags:
        - Bookings
      security:
        - bearerAuth: [ ]
      parameters:
        - in: path
          name: id_user
          required: true
          schema:
            type: string
          description: UUID of the user creating the booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                date:
                  type: string
                  format: date
                  example: "2025-05-22"
                time:
                  type: string
                  example: "14:00"
                serviceId:
                  type: string
                  example: "service_uuid"
                barberId:
                  type: string
                  example: "barber_uuid"
                status:
                  type: string
                  example: "scheduled"
                paymentType:
                  type: string
                  example: "pix"
                nrPrice:
                  type: number
                  example: 50.2
      responses:
        "201":
          description: Booking created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Booking created successfully"
                  bookingId:
                    type: string
                    example: "booking_uuid"
        "500":
          description: Server error

  /booking/query/{id_user}:
    get:
      summary: Get all bookings for a user
      tags:
        - Bookings
      security:
        - bearerAuth: [ ]
      parameters:
        - in: path
          name: id_user
          required: true
          schema:
            type: string
          description: UUID of the user
      responses:
        '200':
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id_booking:
                      type: string
                      example: booking_uuid
                    dt_date:
                      type: string
                      format: date
                      example: 2025-05-22
                    tx_time:
                      type: string
                      example: 14:00
                    service:
                      type: string
                      example: Corte de Cabelo
                    barber_name:
                      type: string
                      example: José Hernandez
                    fl_status:
                      type: string
                      example: scheduled
                    paymentType:
                      type: string
                      example: pix
                    nrPrice:
                      type: number
                      example: 50.2
        '401':
          description: Unauthorized

  /booking/update/{id_user}:
    put:
      summary: Update a booking
      description: Updates the details of an existing booking.
      tags:
        - Bookings
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id_user
          in: path
          required: true
          description: The ID of the user
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bookingId:
                  type: string
                  description: ID of the booking
                  example: "abc123"
                status:
                  type: string
                  description: The status to update
                  example: "completed"
              required:
                - bookingId
                - status
      responses:
        '200':
          description: Booking updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Booking updated successfully
                  booking:
                    type: object
                    properties:
                      bookingId:
                        type: string
                        example: "booking789"
                      status:
                        type: string
                        example: "Completed"
        '400':
          description: Invalid request
        '404':
          description: Booking not found
        '500':
          description: Internal server error

  /barber/query:
    get:
      summary: Lists the barbers
      tags:
        - Barbers
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: Barber list returned successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      example: d7a1f13a-5b32-4f68-8d7a-1c9e4a4f4f01
                    name:
                      type: string
                      example: José Hernandez
                    specialty:
                      type: string
                      example: Fade Specialist
                    img:
                      type: string
                      example: /images/barber01.avif
        '500':
          description: Server error

  /user/query/{id_user}:
    get:
      summary: Get user by ID
      description: Retrieve details of a user by their unique ID.
      tags:
        - User
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id_user
          in: path
          required: true
          description: The ID of the user to retrieve
          schema:
            type: string
            example: "user123"
      responses:
        '200':
          description: User retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: "82d384b6-3d68-49d0-b990-103be3838bb4"
                  name:
                    type: string
                    nullable: true
                    example: null
                  email:
                    type: string
                    example: "matheus@gmail.com"
                  phone:
                    type: string
                    nullable: true
                    example: null
                  birthDate:
                    type: string
                    format: date
                    nullable: true
                    example: null
                  flStatus:
                    type: string
                    description: User status (A = Active, I = Inactive)
                    example: "I"
                  createdAt:
                    type: string
                    format: date-time
                    example: "Fri, 23 May 2025 00:00:00 GMT"
                  updatedAt:
                    type: string
                    format: date-time
                    nullable: true
                    example: null
                  avaliableServicesNumber:
                    type: integer
                    nullable: true
                    example: null
                  loyaltyPackage:
                    type: string
                    nullable: true
                    example: null
        '404':
          description: User not found
        '500':
          description: Internal server error

  /user/update/{id_user}:
    put:
      summary: Update user
      description: Update user information by ID.
      tags:
        - User
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id_user
          in: path
          required: true
          description: The ID of the user to update
          schema:
            type: string
            example: "d6957f05-117f-4d4e-bb22-a98512898ef1"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                  example: "d6957f05-117f-4d4e-bb22-a98512898ef1"
                name:
                  type: string
                  example: "Matheus S Tavares"
                email:
                  type: string
                  example: "matheus1@emial.com"
                phone:
                  type: string
                  example: "+55 79 98863-1138"
                birthDate:
                  type: string
                  format: date
                  example: "1996-03-26"
                avaliableServicesNumber:
                  type: integer
                  example: 0
                loyaltyPackage:
                  type: string
                  example: "monthly"
                createdAt:
                  type: string
                  format: date-time
                  example: "Wed, 21 May 2025 00:00:00 GMT"
              required:
                - id
                - name
                - email
                - phone
                - birthDate
                - avaliableServicesNumber
                - loyaltyPackage
                - createdAt
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Usuário atualizado com sucesso"
        '400':
          description: Invalid request
        '404':
          description: User not found
        '500':
          description: Internal server error

  /activate/{activation_token}:
    post:
      summary: Activate user
      tags:
        - Authentication
      parameters:
        - in: path
          name: activation_token
          required: true
          schema:
            type: string
          description: Activation user token
          example: "f65e7c3a-1a2b-4e5c-9d8f-987654321abc"
      responses:
        '200':
          description: Successful account activation page
          content:
            text/html:
              schema:
                type: string
                example: "<html><body><h1>Your account has been successfully activated.</h1></body></html>"
        '400':
          description: Invalid or missing token
          content:
            text/html:
              schema:
                type: string
                example: "<html><body><h1>Invalid or expired token.</h1></body></html>"
        '404':
          description: User not found
          content:
            text/html:
              schema:
                type: string
                example: "<html><body><h1>User not found.</h1></body></html>"
        '500':
          description: Internal server error
          content:
            text/html:
              schema:
                type: string
                example: "<html><body><h1>Internal server error.</h1></body></html>"

  /service/query:
    get:
      summary: Get all available services grouped by category
      tags:
        - Services
      responses:
        '200':
          description: List of services grouped by category
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    title:
                      type: string
                      example: Haircuts
                    services:
                      type: array
                      items:
                        type: object
                        properties:
                          id:
                            type: string
                            format: uuid
                            example: a1f5d8b2-1e0d-4e3c-b8d1-90f1b8a1c001
                          name:
                            type: string
                            example: Classic Haircut
                          description:
                            type: string
                            example: A traditional cut using clippers or scissors.
                          price:
                            type: number
                            format: float
                            example: 25.0
        '404':
          description: No services found
        '500':
          description: Internal server error